- Changed uint to uint8 to store durationInDays and appealPeriodInDays values

+++ contracts/BrehonContract.sol
@@ -45,7 +45,7 @@ contract BrehonContract is
   mapping (address => uint) public awards;
   Resolution public proposedSettlement;
 
-  uint public appealPeriodInDays = 5;
+  uint8 public appealPeriodInDays = 5;
   uint public appealPeriodStartTime;
 
   event ExecutionStarted(address _caller, uint _totalDeposits);
